Battle Bingo Version 0.1

+++ include/data.h
@@ -83,7 +83,7 @@ struct TrainerMon
     u8 padding1:1;
     u8 dynamaxLevel:4;
     u8 padding2:4;
-    u32 tags;
+    u64 tags;
 };
 
 #define TRAINER_PARTY(partyArray) partyArray, .partySize = ARRAY_COUNT(partyArray)
