+++ src/trainer_pools.c
@@ -17,7 +17,7 @@ static void HasRequiredTag(const struct Trainer *trainer, u8* poolIndexArray, st
     for (u32 currTag = 2; currTag < POOL_NUM_TAGS; currTag++)
     {
         if (rules->tagRequired[currTag]
-         && trainer->party[poolIndexArray[currIndex]].tags & (1u << currTag))
+         && trainer->party[poolIndexArray[currIndex]].tags & (1ULL << currTag))
         {
             *arrayIndex = currIndex;
             *foundRequiredTag = TRUE;
@@ -40,7 +40,7 @@ static u32 DefaultLeadPickFunction(const struct Trainer *trainer, u8 *poolIndexA
         for (u32 currIndex = 0; currIndex < trainer->poolSize; currIndex++)
         {
             if ((poolIndexArray[currIndex] != POOL_SLOT_DISABLED)
-             && (trainer->party[poolIndexArray[currIndex]].tags & (1u << POOL_TAG_LEAD)))
+             && (trainer->party[poolIndexArray[currIndex]].tags & (1ULL << POOL_TAG_LEAD)))
             {
                 if (firstLeadIndex == POOL_SLOT_DISABLED)
                     firstLeadIndex = currIndex;
@@ -79,7 +79,7 @@ static u32 DefaultAcePickFunction(const struct Trainer *trainer, u8 *poolIndexAr
         for (u32 currIndex = 0; currIndex < trainer->poolSize; currIndex++)
         {
             if ((poolIndexArray[currIndex] != POOL_SLOT_DISABLED)
-             && (trainer->party[poolIndexArray[currIndex]].tags & (1u << POOL_TAG_ACE)))
+             && (trainer->party[poolIndexArray[currIndex]].tags & (1ULL << POOL_TAG_ACE)))
             {
                 if (firstAceIndex == POOL_SLOT_DISABLED)
                     firstAceIndex = currIndex;
@@ -114,8 +114,8 @@ static u32 DefaultOtherPickFunction(const struct Trainer *trainer, u8 *poolIndex
     for (u32 currIndex = 0; currIndex < trainer->poolSize; currIndex++)
     {
         if (poolIndexArray[currIndex] != POOL_SLOT_DISABLED
-         && !(trainer->party[poolIndexArray[currIndex]].tags & (1u << POOL_TAG_LEAD))
-         && !(trainer->party[poolIndexArray[currIndex]].tags & (1u << POOL_TAG_ACE)))
+         && !(trainer->party[poolIndexArray[currIndex]].tags & (1ULL << POOL_TAG_LEAD))
+         && !(trainer->party[poolIndexArray[currIndex]].tags & (1ULL << POOL_TAG_ACE)))
         {
             if (firstUnpickedIndex == POOL_SLOT_DISABLED)
                 firstUnpickedIndex = currIndex;
@@ -172,7 +172,7 @@ static u32 PickMonFromPool(const struct Trainer *trainer, u8 *poolIndexArray, u3
     if (monIndex == POOL_SLOT_DISABLED)
         return monIndex;
 
-    u32 chosenTags = trainer->party[monIndex].tags;
+    u64 chosenTags = trainer->party[monIndex].tags;
     enum Species chosenSpecies = trainer->party[monIndex].species;
     enum Item chosenItem = trainer->party[monIndex].heldItem;
     enum NationalDexOrder chosenNatDex = gSpeciesInfo[chosenSpecies].natDexNum;
@@ -200,7 +200,7 @@ static u32 PickMonFromPool(const struct Trainer *trainer, u8 *poolIndexArray, u3
     {
         if (poolIndexArray[currIndex] != POOL_SLOT_DISABLED)
         {
-            u32 currentTags = trainer->party[poolIndexArray[currIndex]].tags;
+            u64 currentTags = trainer->party[poolIndexArray[currIndex]].tags;
             enum Species currentSpecies = trainer->party[poolIndexArray[currIndex]].species;
             enum Item currentItem = trainer->party[poolIndexArray[currIndex]].heldItem;
             enum NationalDexOrder currentNatDex = gSpeciesInfo[currentSpecies].natDexNum;
@@ -343,7 +343,7 @@ static void TestPrune(const struct Trainer *trainer, u8 *poolIndexArray, const s
 
 static void RandomTagPrune(const struct Trainer *trainer, u8 *poolIndexArray, const struct PoolRules *rules)
 {
-    u32 tagToUse = trainer->party[poolIndexArray[0]].tags;
+    u64 tagToUse = trainer->party[poolIndexArray[0]].tags;
     for (u32 i = 0; i < trainer->poolSize; i++)
         if (!(trainer->party[poolIndexArray[i]].tags & tagToUse))
             poolIndexArray[i] = POOL_SLOT_DISABLED;
